comparer
extra material (to be deleted later) to test the correct functioning of cnc. In particular, comparer folder includes two python files, "sorter.py" and "comparer..py", which respectively sort the textual content of a file deleting also the empty lines, and check if the textual contents of different files differ.  The aim of this extra part of the code was to check whether apparently different ttl files generated in test runs of cnc for NOCI contained the very same information, after a structural reorganization.