12_mapping.py update
required changes in order to make any of the functions of the test independent. The input material was triplicated and also the output were splitted, so that each of the functions which modifies the input material does it without affecting the others. In this way it should be easier to check the correct functioning of each one.